Opt-in to the graph statistics, and to opening `dot`
